requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
fastapi==0.110.2
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
orjson==3.10.1
packaging==24.0
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.4
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240425
types-redis==4.6.0.20240425
types-setuptools==69.5.0.20240423
typing_extensions==4.11.0
uvicorn==0.29.0
